Update and rename kalinethunter to kalinethunter.sh
This should be a better standardization of the way script prints information for the user, the coding standard amongst the files being created, and how the variables are used across each file. I also added some declarations for variables in the startkali script to ensure the original variables would remain the same, even if someone mistakenly changes them between logging on to kali-nethunter.